refactor: Convert Exercise1 to separate package structure
Move Exercise1 (Doobie integration) to its own package in chap04/exercise1/:
- exercises/chap04/exercise1/Exercise1.scala (package zionomicon.exercises.chap04.exercise1)
- solutions/chap04/exercise1/Exercise1.scala (package zionomicon.solutions.chap04.exercise1)

This allows the Main object to be directly runnable in IntelliJ IDEA.

Keep Exercise2 (Lepus integration) in the main 04-essentials-integrating-with-zio.scala
file for now.
